pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.4.0"
description = "difference calculator"
authors = ["Metaller <metaller7440326@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
flake8 = "^3.8.4"
pytest = "^6.2.2"
PyYAML = "^5.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"